requirements.txt
anyio==3.7.1
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
djangorestframework==3.15.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.8
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
rfc3986==1.5.0
setuptools==72.1.0
sniffio==1.3.1
sqlparse==0.5.1
urllib3==2.2.2
wheel==0.44.0
